requirements.txt
# Por completar lista de dependencias
numpy
pydantic
python-docx
python-dotenv
rich

uvicorn
fastapi

openai
redis
requests
langchain
langchain-core
langchain-text-splitters
langchain-openai
langchain-community
langsmith
